requirements.txt
chromadb==1.0.17
langchain==0.3.27
langchain_core==0.3.74
langchain_gigachat==0.3.12
langchain_google_genai==2.1.9
langchain_groq==0.3.7
langchain_huggingface==0.3.1
langchain_openai==0.3.30
python-dotenv==1.1.1
Requests==2.32.4
sentence_transformers==5.0.0
zxcvbn==4.5.0
